Módulo para compressão e conversão de imagens #6
Implementada a conversão de imagens para o formato jpg e separação da conversão para um módulo separado (src/utils/images.js).
